# Moves more func inst fields to comptime objects