--- a/internal/engine/compiler/engine.go
+++ b/internal/engine/compiler/engine.go
@@ -284,6 +284,7 @@ type (
 		sourceModule *wasm.Module
 		// listener holds a listener to notify when this function is called.
 		listener experimental.FunctionListener
+		goFunc   interface{}
 
 		sourceOffsetMap *sourceOffsetMap
 	}
@@ -352,7 +353,7 @@ const (
 	tableInstanceTableLenOffset = 8
 
 	// Offsets for wasm.FunctionInstance.
-	functionInstanceTypeIDOffset = 40
+	functionInstanceTypeIDOffset = 16
 
 	// Offsets for wasm.MemoryInstance.
 	memoryInstanceBufferOffset    = 0
@@ -496,6 +497,7 @@ func (e *engine) CompileModule(ctx context.Context, module *wasm.Module, listene
 		return err
 	}
 
+	var withGoFunc bool
 	importedFuncs := module.ImportFuncCount()
 	funcs := make([]*code, len(module.FunctionSection))
 	ln := len(listeners)
@@ -509,10 +511,12 @@ func (e *engine) CompileModule(ctx context.Context, module *wasm.Module, listene
 		funcIndex := wasm.Index(i)
 		var compiled *code
 		if ir.GoFunc != nil {
+			withGoFunc = true
 			if compiled, err = compileGoDefinedHostFunction(cmp); err != nil {
 				def := module.FunctionDefinitionSection[funcIndex+importedFuncs]
 				return fmt.Errorf("error compiling host go func[%s]: %w", def.DebugName(), err)
 			}
+			compiled.goFunc = ir.GoFunc
 		} else if compiled, err = compileWasmFunction(cmp, ir); err != nil {
 			def := module.FunctionDefinitionSection[funcIndex+importedFuncs]
 			return fmt.Errorf("error compiling wasm func[%s]: %w", def.DebugName(), err)
@@ -526,7 +530,7 @@ func (e *engine) CompileModule(ctx context.Context, module *wasm.Module, listene
 		compiled.sourceModule = module
 		funcs[funcIndex] = compiled
 	}
-	return e.addCodes(module, funcs)
+	return e.addCodes(module, funcs, withGoFunc)
 }
 
 // NewModuleEngine implements the same method as documented on wasm.Engine.
@@ -908,7 +912,7 @@ entry:
 			}
 			stack := ce.stack[base : base+stackLen]
 
-			fn := calleeHostFunction.source.GoFunc
+			fn := calleeHostFunction.parent.goFunc
 			switch fn := fn.(type) {
 			case api.GoModuleFunction:
 				fn.Call(ce.ctx, callCtx.WithMemory(ce.memoryInstance), stack)