@mathetake 2022-12-21
closing for now as this is incompatible with the namespace stuff