@mathetake 2023-01-09
Revived this PR since I've come up withe the solution... and here's the bench result (1KB reduction in instantiation)
```
$ benchstat old.txt new.txt
name                                    old time/op    new time/op    delta
Initialization/interpreter-10             19.8µs ± 5%    20.3µs ±11%    ~     (p=0.400 n=9+10)
Initialization/interpreter-multiple-10    11.7µs ± 8%    12.7µs ±18%    ~     (p=0.052 n=10+10)
Initialization/compiler-10                19.8µs ±61%    20.2µs ±55%    ~     (p=0.912 n=10+10)
Initialization/compiler-multiple-10       11.7µs ± 4%    11.8µs ±11%    ~     (p=0.853 n=10+10)

name                                    old alloc/op   new alloc/op   delta
Initialization/interpreter-10              138kB ± 0%     136kB ± 0%  -0.93%  (p=0.000 n=10+10)
Initialization/interpreter-multiple-10     138kB ± 0%     137kB ± 0%  -0.91%  (p=0.000 n=8+9)
Initialization/compiler-10                 142kB ± 0%     141kB ± 0%  -0.63%  (p=0.002 n=8+10)
Initialization/compiler-multiple-10        143kB ± 0%     142kB ± 0%  -0.62%  (p=0.000 n=8+10)

name                                    old allocs/op  new allocs/op  delta
Initialization/interpreter-10               52.0 ± 0%      52.0 ± 0%    ~     (all equal)
Initialization/interpreter-multiple-10      58.0 ± 0%      58.0 ± 0%    ~     (all equal)
Initialization/compiler-10                  42.0 ± 0%      42.0 ± 0%    ~     (all equal)
Initialization/compiler-multiple-10         48.0 ± 0%      48.0 ± 0%    ~     (all equal)

```